% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newplots.R
\name{stat_lm}
\alias{stat_lm}
\alias{geom_lm}
\title{Linear Model Displays}
\usage{
stat_lm(mapping = NULL, data = NULL, geom = "lm", position = "identity",
  interval = c("none", "prediction", "confidence"), level = 0.95,
  formula = y ~ x, lm.args = list(), backtrans = identity, ...,
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)

geom_lm(mapping = NULL, data = NULL, stat = "lm", position = "identity",
  interval = c("none", "prediction", "confidence"), level = 0.95,
  formula = y ~ x, lm.args = list(), backtrans = identity, ...,
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{geom, stat}{Use to override the default connection between
`geom_lm` and `stat_lm`.}

\item{interval}{One of `"none"`, `"confidence"` or `"prediction"`.}

\item{level}{The level used for confidence or prediction intervals}

\item{formula}{a formula describing the model in terms of `y` (response)
and `x` (predictor).}

\item{backtrans}{a function that transforms the response back to
the original scale when the `formula` includes a transforamtion on
`y`.}
}
\description{
Adds linear model fits to plots. `geom_lm()` and `stat_lm()` are essentially
equivalent.  Use `geom_lm()` unless you want a non-standard geom.
}
\details{
Stat calculation is performed by the (currently undocumented)
`predictdf`.  Pointwise confidence or prediction bands are
calculated using the [predict()] method.
}
\section{Aesthetics}{

\aesthetics{geom}{lm}
}

\examples{
if (require(mosaicData)) {
  ggplot(data = KidsFeet, aes(y = length, x = width, color = sex)) +
    geom_lm() +
    geom_point()
  ggplot(data = KidsFeet, aes(y = length, x = width, color = sex)) +
    geom_lm(interval = "prediction", color = "skyblue") +
    geom_lm(interval = "confidence") +
    geom_point() +
    facet_wrap(~sex)
 # non-standard display
  ggplot(data = KidsFeet, aes(y = length, x = width, color = sex)) +
    stat_lm(aes(fill = sex), color = NA, interval = "confidence", geom = "ribbon",
    alpha = 0.2) +
    geom_point() +
    facet_wrap(~sex)
ggplot(mpg, aes(displ, hwy)) +
   geom_lm(formula = log(y) ~ poly(x,3), backtrans = exp,
     interval = "prediction", fill = "skyblue") +
   geom_lm(formula = log(y) ~ poly(x,3), backtrans = exp, interval = "confidence",
     color = "red") +
   geom_point()
}
}
\seealso{
[lm()] for details on linear model fitting.
}
