% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{get_labels}
\alias{get_labels}
\alias{get_labels.data.frame}
\alias{get_labels.labelled}
\alias{get_labels.default}
\alias{set_labels}
\alias{set_labels.default}
\alias{set_labels.data.frame}
\alias{print.gf_ggplot}
\title{Set and extract labels from a labeled object}
\usage{
get_labels(object, ...)

\method{get_labels}{data.frame}(object, ...)

\method{get_labels}{labelled}(object, ...)

\method{get_labels}{default}(object, ...)

set_labels(object, ...)

\method{set_labels}{default}(object, lab = NULL, ...)

\method{set_labels}{data.frame}(object, lab, ...)

\method{print}{gf_ggplot}(x, lab = get_labels(x$data), ...)
}
\arguments{
\item{object}{An R object, potentially labelled.}

\item{...}{Additional arguments. Ignored in existing methods.}

\item{lab}{A string providing a label (when \code{object} is a vector, for example)
or a named list of strings
providing multiple labels (when \code{object} is a data frame, for example).}

\item{x}{A ggplot.}
}
\value{
The label(s) for the object.  If \code{object} is a data frame,
labels are extracted from each variable and returned as a named list.
If there are no labels, \code{NULL} is returned.
}
\description{
Some packages like expss provide mechanisms for providing longer labels to R objects.
These labels can be used when labeling plots and tables, for example, without requiring
long or awkward variable names.  This is an experimental feature and currently only supports
expss or any other system that stores a label in the \code{label} attribute of a vector.
}
\examples{
KF <-
  mosaicData::KidsFeet \%>\%
  set_labels(
    list(
      length      = 'foot length (cm)',
      width       = 'foot width (cm)',
      birthmonth  = 'birth month',
      birthyear   = 'birth year',
      biggerfoot  = 'bigger foot',
      domhand     = 'dominant hand'
    )
  )
KF \%>\%
  gf_point(length ~ width, color = ~ domhand)
get_labels(KF)


}
