% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_aux.R
\name{gf_labs}
\alias{gf_labs}
\alias{gf_lims}
\alias{gf_theme}
\alias{gf_facet_wrap}
\alias{gf_facet_grid}
\alias{gf_refine}
\title{Non-layer functions for gf plots}
\usage{
gf_labs(object, ...)

gf_lims(object, ...)

gf_theme(object, ..., theme = theme_minimal)

gf_facet_wrap(object, ...)

gf_facet_grid(object, ...)

gf_refine(object, ...)
}
\arguments{
\item{object}{a gg object}

\item{...}{additional arguments passed through to the similarly named function in
\pkg{ggplot2}.}

\item{theme}{a ggplot2 theme function like \code{\link{theme_minimal}}.
ggplot2}
}
\value{
a modified gg object
}
\description{
These functions modify things like labels, limits, scales, etc. for plots
ggplot2 plots. They are wrappers around functions in ggplot2 that allow for
chaining syntax.
}
