% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiraph-documentation-with-examples.R
\name{gf_bar_interactive}
\alias{gf_bar_interactive}
\title{Interactive bar charts}
\usage{
gf_bar_interactive(
  object = NULL,
  gformula = NULL,
  data = NULL,
  ...,
  alpha,
  color,
  fill,
  group,
  linetype,
  linewidth,
  width = NULL,
  xlab,
  ylab,
  title,
  subtitle,
  caption,
  stat = "count",
  position = "stack",
  show.legend = NA,
  show.help = NULL,
  inherit = TRUE,
  environment = parent.frame()
)
}
\arguments{
\item{object}{When chaining, this holds an object produced in the
earlier portions of the chain. Most users can safely ignore this
argument. See details and examples.}

\item{gformula}{A formula, typically with shape \code{~ x}. (\code{y ~
x} is also possible, but typically using one of
\code{\link[=gf_col]{gf_col()}}, \code{\link[=gf_props]{gf_props()}},
or \code{\link[=gf_percents]{gf_percents()}} is preferable to using
this formula shape.) Faceting can be achieved by including \code{|} in
the formula.}

\item{data}{The data to be displayed in this layer. There are three
options: If \code{NULL}, the default, the data is inherited from the
plot data as specified in the call to
\code{\link[ggplot2:ggplot]{ggplot()}}. A \code{data.frame}, or other
object, will override the plot data. All objects will be fortified to
produce a data frame. See \code{\link[ggplot2:fortify]{fortify()}} for
which variables will be created. A \code{function} will be called with
a single argument, the plot data. The return value must be a
\code{data.frame}, and will be used as the layer data. A
\code{function} can be created from a \code{formula} (e.g. \code{~
head(.x, 10)}).}

\item{...}{Additional arguments passed to the underlying interactive
geom. This is where ggiraph's interactive aesthetics are supplied,
including \code{tooltip} (text shown on hover), \code{data_id}
(identifiers used for interactive selection), and \code{onclick}
(JavaScript run on click).}

\item{alpha}{Opacity (0 = invisible, 1 = opaque).}

\item{color}{A color or a formula used for mapping color.}

\item{fill}{A color for filling, or a formula used for mapping fill.}

\item{group}{Used for grouping.}

\item{linetype}{A linetype (numeric or "dashed", "dotted", etc.) or a
formula used for mapping linetype.}

\item{linewidth}{A numerical line width or a formula used for mapping
linewidth.}

\item{width}{Width of the bars.}

\item{xlab}{Label for x-axis. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{ylab}{Label for y-axis. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{title}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{subtitle}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{caption}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{stat}{Override the default connection between \code{geom_bar()}
and \code{stat_count()}. For more information about overriding these
connections, see how the \link[ggplot2:layer_stats]{stat} and
\link[ggplot2:layer_geoms]{geom} arguments work.}

\item{position}{A position adjustment to use on the data for this layer.
This can be used in various ways, including to prevent overplotting
and improving the display. The \code{position} argument accepts the
following: \itemize{ \item The result of calling a position function,
such as \code{position_jitter()}. This method allows for passing extra
arguments to the position. \item A string naming the position
adjustment. To give the position as a string, strip the function name
of the \code{position_} prefix. For example, to use
\code{position_jitter()}, give the position as \code{"jitter"}. \item
For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation. }}

\item{show.legend}{logical. Should this layer be included in the
legends? \code{NA}, the default, includes if any aesthetics are
mapped. \code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics
to display. To include legend keys for all levels, even when no data
exists, use \code{TRUE}. If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{show.help}{If \code{TRUE}, display some minimal help.}

\item{inherit}{A logical indicating whether default attributes are
inherited.}

\item{environment}{An environment in which to look for variables not
found in \code{data}.}
}
\value{
A gg object that can be displayed with \code{\link[=gf_girafe]{gf_girafe()}}.
}
\description{
Creates an interactive plot using ggiraph. This function extends
\code{\link[=gf_bar]{gf_bar()}} with interactive features like tooltips and clickable elements.
}
\section{Additional interactive features}{

\itemize{
\item \code{onclick}: JavaScript code (as character string) executed when clicking elements.
\item Additional ggiraph aesthetics may be available depending on the geom.
}
}

\examples{

diamonds |>
  gf_bar_interactive(
    ~color,
    fill = ~cut,
    tooltip = ~ stage(
      start = glue::glue("color: {color}; cut: {cut}"),
      after_stat = glue::glue("{tooltip}; count = {count}")
    ),
    data_id = ~ glue::glue("{cut} -- {color}"),
    size = 3
  ) |>
  gf_girafe()

}
\seealso{
\code{\link[=gf_bar]{gf_bar()}}, \code{\link[=gf_girafe]{gf_girafe()}}
}
