% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiraph-documentation-with-examples.R
\name{gf_crossbar_interactive}
\alias{gf_crossbar_interactive}
\title{Interactive crossbar plots}
\usage{
gf_crossbar_interactive(
  object = NULL,
  gformula = NULL,
  data = NULL,
  ...,
  alpha,
  color,
  group,
  middle.linetype,
  middle.colour = NULL,
  middle.color = NULL,
  middle.linewidth = NULL,
  box.colour = NULL,
  box.color = NULL,
  box.linetype = NULL,
  box.linewidth = NULL,
  na.rm = FALSE,
  inherit.aes = TRUE,
  xlab,
  ylab,
  title,
  subtitle,
  caption,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  show.help = NULL,
  inherit = TRUE,
  environment = parent.frame()
)
}
\arguments{
\item{object}{When chaining, this holds an object produced in the
earlier portions of the chain. Most users can safely ignore this
argument. See details and examples.}

\item{gformula}{A formula with shape \code{y + ymin + ymax ~ x}.
Faceting can be achieved by including \code{|} in the formula.}

\item{data}{The data to be displayed in this layer. There are three
options: If \code{NULL}, the default, the data is inherited from the
plot data as specified in the call to
\code{\link[ggplot2:ggplot]{ggplot()}}. A \code{data.frame}, or other
object, will override the plot data. All objects will be fortified to
produce a data frame. See \code{\link[ggplot2:fortify]{fortify()}} for
which variables will be created. A \code{function} will be called with
a single argument, the plot data. The return value must be a
\code{data.frame}, and will be used as the layer data. A
\code{function} can be created from a \code{formula} (e.g. \code{~
head(.x, 10)}).}

\item{...}{Additional arguments passed to the underlying interactive
geom. This is where ggiraph's interactive aesthetics are supplied,
including \code{tooltip} (text shown on hover), \code{data_id}
(identifiers used for interactive selection), and \code{onclick}
(JavaScript run on click).}

\item{alpha}{Opacity (0 = invisible, 1 = opaque).}

\item{color}{A color or a formula used for mapping color.}

\item{group}{Used for grouping.}

\item{middle.linetype}{Arguments to control the middle line.}

\item{middle.colour}{Arguments to control the middle line.}

\item{middle.color}{Arguments to control the middle line.}

\item{middle.linewidth}{Arguments to control the middle line.}

\item{box.colour}{Arguments to control the box.}

\item{box.color}{Arguments to control the box.}

\item{box.linetype}{Arguments to control the box.}

\item{box.linewidth}{Arguments to control the box.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed
with a warning. If \code{TRUE}, missing values are silently removed.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper
functions that define both data and aesthetics and shouldn't inherit
behaviour from the default plot specification, e.g.
\code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{xlab}{Label for x-axis. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{ylab}{Label for y-axis. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{title}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{subtitle}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{caption}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{stat}{The statistical transformation to use on the data for this
layer. When using a \verb{geom_*()} function to construct a layer, the
\code{stat} argument can be used to override the default coupling
between geoms and stats. The \code{stat} argument accepts the
following: \itemize{ \item A \code{Stat} ggproto subclass, for example
\code{StatCount}. \item A string naming the stat. To give the stat as
a string, strip the function name of the \code{stat_} prefix. For
example, to use \code{stat_count()}, give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation. }}

\item{position}{A position adjustment to use on the data for this layer.
This can be used in various ways, including to prevent overplotting
and improving the display. The \code{position} argument accepts the
following: \itemize{ \item The result of calling a position function,
such as \code{position_jitter()}. This method allows for passing extra
arguments to the position. \item A string naming the position
adjustment. To give the position as a string, strip the function name
of the \code{position_} prefix. For example, to use
\code{position_jitter()}, give the position as \code{"jitter"}. \item
For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation. }}

\item{show.legend}{logical. Should this layer be included in the
legends? \code{NA}, the default, includes if any aesthetics are
mapped. \code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics
to display. To include legend keys for all levels, even when no data
exists, use \code{TRUE}. If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{show.help}{If \code{TRUE}, display some minimal help.}

\item{inherit}{A logical indicating whether default attributes are
inherited.}

\item{environment}{An environment in which to look for variables not
found in \code{data}.}
}
\value{
A gg object that can be displayed with \code{\link[=gf_girafe]{gf_girafe()}}.
}
\description{
Creates an interactive plot using ggiraph. This function extends
\code{\link[=gf_crossbar]{gf_crossbar()}} with interactive features like tooltips and clickable elements.
}
\section{Additional interactive features}{

\itemize{
\item \code{onclick}: JavaScript code (as character string) executed when clicking elements.
\item Additional ggiraph aesthetics may be available depending on the geom.
}
}

\examples{

diamonds |>
  dplyr::filter(carat < 1.1, carat > 0.9) |>
  dplyr::group_by(color, cut) |>
  dplyr::summarise(
    median_price = median(price) |> round(),
    lower = quantile(price, 0.25) |> round(),
    upper = quantile(price, 0.75) |> round(),
    iqr = upper - lower
  ) |>
  gf_crossbar_interactive(
    cut ~ median_price + lower + upper | color,
    color = ~ cut,
    tooltip = ~ paste0(
      "75th percentile: ", upper,
      "\nmedian: ", median_price,
      "\n25th percentile: ", lower
      )
  ) |>
  gf_girafe()

}
\seealso{
\code{\link[=gf_crossbar]{gf_crossbar()}}, \code{\link[=gf_girafe]{gf_girafe()}}
}
