% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiraph-documentation-with-examples.R
\name{gf_curve_interactive}
\alias{gf_curve_interactive}
\alias{gf_segment_interactive}
\title{Interactive curve plots}
\usage{
gf_curve_interactive(
  object = NULL,
  gformula = NULL,
  data = NULL,
  ...,
  alpha,
  color,
  group,
  linetype,
  linewidth,
  curvature = 0.5,
  angle = 90,
  ncp = 5,
  arrow = NULL,
  lineend = "butt",
  xlab,
  ylab,
  title,
  subtitle,
  caption,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  show.help = NULL,
  inherit = TRUE,
  environment = parent.frame()
)

gf_segment_interactive(
  object = NULL,
  gformula = NULL,
  data = NULL,
  ...,
  alpha,
  color,
  group,
  linetype,
  linewidth,
  arrow = NULL,
  lineend = "butt",
  xlab,
  ylab,
  title,
  subtitle,
  caption,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  show.help = NULL,
  inherit = TRUE,
  environment = parent.frame()
)
}
\arguments{
\item{object}{When chaining, this holds an object produced in the
earlier portions of the chain. Most users can safely ignore this
argument. See details and examples.}

\item{gformula}{A formula with shape \code{y + yend ~ x + xend}.}

\item{data}{The data to be displayed in this layer. There are three
options: If \code{NULL}, the default, the data is inherited from the
plot data as specified in the call to
\code{\link[ggplot2:ggplot]{ggplot()}}. A \code{data.frame}, or other
object, will override the plot data. All objects will be fortified to
produce a data frame. See \code{\link[ggplot2:fortify]{fortify()}} for
which variables will be created. A \code{function} will be called with
a single argument, the plot data. The return value must be a
\code{data.frame}, and will be used as the layer data. A
\code{function} can be created from a \code{formula} (e.g. \code{~
head(.x, 10)}).}

\item{...}{Additional arguments passed to the underlying interactive
geom. This is where ggiraph's interactive aesthetics are supplied,
including \code{tooltip} (text shown on hover), \code{data_id}
(identifiers used for interactive selection), and \code{onclick}
(JavaScript run on click).}

\item{alpha}{Opacity (0 = invisible, 1 = opaque).}

\item{color}{A color or a formula used for mapping color.}

\item{group}{Used for grouping.}

\item{linetype}{A linetype (numeric or "dashed", "dotted", etc.) or a
formula used for mapping linetype.}

\item{linewidth}{A numerical line width or a formula used for mapping
linewidth.}

\item{curvature}{A numeric value giving the amount of curvature.
Negative values produce left-hand curves, positive values produce
right-hand curves, and zero produces a straight line.}

\item{angle}{A numeric value between 0 and 180, giving an amount to skew
the control points of the curve. Values less than 90 skew the curve
towards the start point and values greater than 90 skew the curve
towards the end point.}

\item{ncp}{The number of control points used to draw the curve. More
control points creates a smoother curve.}

\item{arrow}{specification for arrow heads, as created by
\code{\link[grid:arrow]{grid::arrow()}}.}

\item{lineend}{Line end style (round, butt, square).}

\item{xlab}{Label for x-axis. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{ylab}{Label for y-axis. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{title}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{subtitle}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{caption}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{stat}{The statistical transformation to use on the data for this
layer. When using a \verb{geom_*()} function to construct a layer, the
\code{stat} argument can be used to override the default coupling
between geoms and stats. The \code{stat} argument accepts the
following: \itemize{ \item A \code{Stat} ggproto subclass, for example
\code{StatCount}. \item A string naming the stat. To give the stat as
a string, strip the function name of the \code{stat_} prefix. For
example, to use \code{stat_count()}, give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation. }}

\item{position}{A position adjustment to use on the data for this layer.
This can be used in various ways, including to prevent overplotting
and improving the display. The \code{position} argument accepts the
following: \itemize{ \item The result of calling a position function,
such as \code{position_jitter()}. This method allows for passing extra
arguments to the position. \item A string naming the position
adjustment. To give the position as a string, strip the function name
of the \code{position_} prefix. For example, to use
\code{position_jitter()}, give the position as \code{"jitter"}. \item
For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation. }}

\item{show.legend}{logical. Should this layer be included in the
legends? \code{NA}, the default, includes if any aesthetics are
mapped. \code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics
to display. To include legend keys for all levels, even when no data
exists, use \code{TRUE}. If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{show.help}{If \code{TRUE}, display some minimal help.}

\item{inherit}{A logical indicating whether default attributes are
inherited.}

\item{environment}{An environment in which to look for variables not
found in \code{data}.}
}
\value{
A gg object that can be displayed with \code{\link[=gf_girafe]{gf_girafe()}}.

A gg object that can be displayed with \code{\link[=gf_girafe]{gf_girafe()}}.
}
\description{
Creates an interactive plot using ggiraph. These functions extend
\code{\link[=gf_segment]{gf_segment()}} and  \code{\link[=gf_curve]{gf_curve()}} with interactive features like tooltips and
clickable elements.

Creates an interactive plot using ggiraph. This function extends
\code{\link[=gf_segment]{gf_segment()}} with interactive features like tooltips and clickable elements.
}
\section{Additional interactive features}{

\itemize{
\item \code{onclick}: JavaScript code (as character string) executed when clicking elements.
\item Additional ggiraph aesthetics may be available depending on the geom.
}


\itemize{
\item \code{onclick}: JavaScript code (as character string) executed when clicking elements.
\item Additional ggiraph aesthetics may be available depending on the geom.
}
}

\examples{
  gf_curve_interactive(
    1 + 2 ~ 0 + 4, color = "red", curvature = - 0.2,
    tooltip = ~ "curvature: -0.2",
    data_id = 0.2
  ) |>
  gf_curve_interactive(
    1 + 2 ~ 0 + 4, color = "blue", curvature = 0.4,
    tooltip = ~ "curvature: 0.4",
    data_id = 0.4) |>
  gf_segment_interactive(
    1 + 2 ~ 0 + 4, color = "green",
    tooltip = ~ "curvature: 0",
    data_id = 0
  ) |>
  gf_girafe(
    options = list(
      opts_hover(css = "stroke: black; stroke-width: 3;", nearest_distance = 10)
    )
  )

}
\seealso{
\code{\link[=gf_curve]{gf_curve()}}, \code{\link[=gf_girafe]{gf_girafe()}}

\code{\link[=gf_segment]{gf_segment()}}, \code{\link[=gf_girafe]{gf_girafe()}}
}
