% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_dist.R
\name{gf_dist}
\alias{gf_dist}
\title{Plot distributions}
\usage{
gf_dist(object = geom_blank(), dist, ..., xmin = NULL, xmax = NULL,
  kind = c("density", "cdf", "qq", "qqstep", "histogram"),
  resolution = 5000L, params = NULL)
}
\arguments{
\item{object}{a gg object.}

\item{dist}{A character string providing the name of a distribution.  Any
distribution for which the functions with names formed by prepending
"d", "p", or "q" to \code{dist} exist can be used.}

\item{...}{additional arguments passed both to the distribution functions and
to the layer.  Note: avoid possible amiguities using \code{params}.}

\item{kind}{One of `"density"`, `"cdf"`, `"qq"`, `"qqstep"`, or `"histogram"`
describing what kind of plot to create.}

\item{resolution}{An integer specifying the number of points to use for creating
the plot.}

\item{params}{a list of parameters for the distribution.}
}
\description{
Create a layer displaying a probability distribution.
}
