% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_functions.R
\name{gf_empty}
\alias{gf_empty}
\title{Create an "empty" plot}
\usage{
gf_empty(environment = parent.frame())
}
\arguments{
\item{environment}{An environment passed to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}}
}
\value{
A plot with now layers.
}
\description{
This is primarily useful as a way to start a sequence of piped
plot layers.
}
\examples{
gf_empty()
data(penguins, package = "palmerpenguins")
gf_empty() \%>\%
  gf_point(bill_length_mm ~ bill_depth_mm, data = penguins, color = ~species)
}
