% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_function_helpers.R
\name{gf_factory}
\alias{gf_factory}
\title{These functions provide a formula interface to \code{ggplot2} and
various geoms. For plots with just one layer, the formula interface
is more compact and is consistent with modeling and mosaic notation.
The functions generate a \code{ggplot} command string which can be displayed by
setting \code{verbose = TRUE} as an argument.}
\usage{
gf_factory(type, aes_form = y ~ x, extras = alist(), note = NULL,
  function_name = paste0("gf_", type))
}
\arguments{
\item{gformula}{A formula describing the x and y variables and other aesthetics in
a form like \code{y ~ x + color:"red" + shape:sex + alpha:0.5}.
The environment of \code{gformula} determines
where the \pkg{ggplot2} code is evaluated.
See details.}

\item{add}{If \code{TRUE} then construct just the layer with no frame.  The result
can be added to an existing frame.}

\item{verbose}{If \code{TRUE} print the ggplot2 command in the console.}

\item{geom}{A way to specify ggplot geoms that are not aliased to gf functions.}

\item{...}{Other arguments such as \code{position="dodge"}.}

\item{show.help}{If \code{TRUE}, display some minimal help.  In particular,
the help will show (a) which geom from \pkg{ggplot2} is used,
(b) how aesthetics are assigned based on \code{formula}, and (c)
any default values of arguments to the geom.}
}
\description{
Formulas must specify the \code{y} and \code{x} aesthetics in the form \code{y ~ x}.
Additional terms of the form \code{attribute::value} map \code{attribute}
to \code{value}.
Additional terms of the form \code{attribute:value} will map \code{attribute}
to \code{value} if \code{value} is the name of a variable in \code{data}, else
\code{attribute} will be set to the constant \code{value}.
Attributes can also be set by including optional arguments of the form
\code{attribute = value}.
}
