% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_functions.R
\name{gf_counts}
\alias{gf_counts}
\alias{gf_bar}
\alias{gf_freqpoly}
\alias{gf_density}
\alias{gf_histogram}
\alias{gf_dotplot}
\alias{gf_dens}
\alias{gf_qq}
\title{Univariate gf_ plotting functions}
\usage{
gf_counts(object = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(object, c("gg", "ggplot")), ...,
  show.help = FALSE)

gf_bar(object = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(object, c("gg", "ggplot")), ...,
  show.help = FALSE)

gf_freqpoly(object = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(object, c("gg", "ggplot")), ...,
  show.help = FALSE)

gf_density(object = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(object, c("gg", "ggplot")), ...,
  show.help = FALSE)

gf_histogram(object = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(object, c("gg", "ggplot")), ...,
  show.help = FALSE)

gf_dotplot(object = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(object, c("gg", "ggplot")), ...,
  show.help = FALSE)

gf_dens(object = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(object, c("gg", "ggplot")), ...,
  show.help = FALSE)

gf_qq(object = NULL, formula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(object, c("gg", "ggplot")), ...,
  show.help = FALSE)
}
\arguments{
\item{object}{When chaining, this holds an object produced in the earlier portions
of the chain.  Most users can safely ignore this argument.
See details and examples.}

\item{formula}{A formula describing the x variable and other aesthetics in
a form like \code{ ~ x + color:red + fill:gray50 + alpha:0.5}}

\item{data}{A data frame with the variables to be plotted}

\item{geom}{A way to specify ggplot geoms that are not aliased to gf functions.}

\item{verbose}{If \code{TRUE} print the ggplot2 command in the console.}

\item{add}{If \code{TRUE} then construct just the layer with no frame.  The result
can be added to an existing frame.}

\item{...}{Other arguments such as \code{position="dodge"}.}

\item{show.help}{If \code{TRUE}, display some minimal help.  In particular,
the help will show (a) which geom from \pkg{ggplot2} is used,
(b) how aesthetics are assigned based on \code{formula}, and (c)
any default values of arguments to the geom.}
}
\description{
These functions provide a formula interface to \code{ggplot2} and
various geoms. The formula interface is similar to the one used
for \pkg{lattice} plots, but more expressive, and consistent with
its use in modeling functions like \code{\link{lm}()}.  These functions
can be used to create a complete plot, or they can be chained together
using the pipe operator from \pkg{magrittr} to create multi-layer plots.
The functions generate a \code{ggplot2} command string which can be
displayed by setting \code{verbose = TRUE} as an argument.
}
\examples{
gf_histogram(~ Sepal.Length + fill:Species, data = iris)
gf_density(~ Sepal.Length + color:Species, data = iris)
gf_dens(~ Sepal.Length + color:Species, data = iris)
gf_freqpoly(~ Sepal.Length + color:Species, data = iris)
gf_dotplot(~ Sepal.Length + fill:Species, data = iris)
gf_counts(~ Species, data = iris)
# Chaining in the data
iris \%>\% gf_dens(~ Sepal.Length + color:Species)
}
\seealso{
\code{\link{gf_point}()}, \code{\link{gf_abline}()}, \code{\link{gf_pointrange}()}, \code{\link{gf_refine}()}, and the other functions documented with these functions.
}
