% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiraph-documentation-with-examples.R
\name{gf_histogram_interactive}
\alias{gf_histogram_interactive}
\title{Interactive histograms}
\usage{
gf_histogram_interactive(
  object = NULL,
  gformula = NULL,
  data = NULL,
  ...,
  bins,
  binwidth,
  alpha = 0.5,
  color,
  fill,
  group,
  linetype,
  linewidth,
  xlab,
  ylab,
  title,
  subtitle,
  caption,
  stat = "bin",
  position = "stack",
  show.legend = NA,
  show.help = NULL,
  inherit = TRUE,
  environment = parent.frame()
)
}
\arguments{
\item{object}{When chaining, this holds an object produced in the
earlier portions of the chain. Most users can safely ignore this
argument. See details and examples.}

\item{gformula}{A formula with shape \code{~ x} (or \code{y ~ x}, but
this shape is not generally needed).}

\item{data}{The data to be displayed in this layer. There are three
options: If \code{NULL}, the default, the data is inherited from the
plot data as specified in the call to
\code{\link[ggplot2:ggplot]{ggplot()}}. A \code{data.frame}, or other
object, will override the plot data. All objects will be fortified to
produce a data frame. See \code{\link[ggplot2:fortify]{fortify()}} for
which variables will be created. A \code{function} will be called with
a single argument, the plot data. The return value must be a
\code{data.frame}, and will be used as the layer data. A
\code{function} can be created from a \code{formula} (e.g. \code{~
head(.x, 10)}).}

\item{...}{Additional arguments passed to the underlying interactive
geom. This is where ggiraph's interactive aesthetics are supplied,
including \code{tooltip} (text shown on hover), \code{data_id}
(identifiers used for interactive selection), and \code{onclick}
(JavaScript run on click).}

\item{bins}{Number of bins. Overridden by \code{binwidth}. Defaults to
30.}

\item{binwidth}{The width of the bins. Can be specified as a numeric
value or as a function that takes x after scale transformation as
input and returns a single numeric value. When specifying a function
along with a grouping structure, the function will be called once per
group. The default is to use the number of bins in \code{bins},
covering the range of the data. You should always override this value,
exploring multiple widths to find the best to illustrate the stories
in your data. The bin width of a date variable is the number of days
in each time; the bin width of a time variable is the number of
seconds.}

\item{alpha}{Opacity (0 = invisible, 1 = opaque).}

\item{color}{A color or a formula used for mapping color.}

\item{fill}{A color for filling, or a formula used for mapping fill.}

\item{group}{Used for grouping.}

\item{linetype}{A linetype (numeric or "dashed", "dotted", etc.) or a
formula used for mapping linetype.}

\item{linewidth}{A numerical line width or a formula used for mapping
linewidth.}

\item{xlab}{Label for x-axis. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{ylab}{Label for y-axis. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{title}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{subtitle}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{caption}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{stat}{Use to override the default connection between
\code{geom_histogram()}/\code{geom_freqpoly()} and \code{stat_bin()}.
For more information at overriding these connections, see how the
\link[ggplot2:layer_stats]{stat} and \link[ggplot2:layer_geoms]{geom}
arguments work.}

\item{position}{A position adjustment to use on the data for this layer.
This can be used in various ways, including to prevent overplotting
and improving the display. The \code{position} argument accepts the
following: \itemize{ \item The result of calling a position function,
such as \code{position_jitter()}. This method allows for passing extra
arguments to the position. \item A string naming the position
adjustment. To give the position as a string, strip the function name
of the \code{position_} prefix. For example, to use
\code{position_jitter()}, give the position as \code{"jitter"}. \item
For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation. }}

\item{show.legend}{logical. Should this layer be included in the
legends? \code{NA}, the default, includes if any aesthetics are
mapped. \code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics
to display. To include legend keys for all levels, even when no data
exists, use \code{TRUE}. If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{show.help}{If \code{TRUE}, display some minimal help.}

\item{inherit}{A logical indicating whether default attributes are
inherited.}

\item{environment}{An environment in which to look for variables not
found in \code{data}.}
}
\value{
A gg object that can be displayed with \code{\link[=gf_girafe]{gf_girafe()}}.
}
\description{
Creates an interactive plot using ggiraph. This function extends
\code{\link[=gf_histogram]{gf_histogram()}} with interactive features like tooltips and clickable elements.
}
\section{Additional interactive features}{

\itemize{
\item \code{onclick}: JavaScript code (as character string) executed when clicking elements.
\item Additional ggiraph aesthetics may be available depending on the geom.
}
}

\examples{
# Interactive histogram with bin information
mtcars |>
  gf_histogram_interactive(
    ~ mpg,
    tooltip = ~ paste0('Min: ', round(after_stat(xmin), 1),
                     '; Max: ', round(after_stat(xmax),1),
                     '; Count: ', after_stat(count)),
    bins = 15) |>
  gf_girafe()

}
\seealso{
\code{\link[=gf_histogram]{gf_histogram()}}, \code{\link[=gf_girafe]{gf_girafe()}}
}
