% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiraph.R
\name{gf_labeller_interactive}
\alias{gf_labeller_interactive}
\title{Create interactive labeller}
\usage{
gf_labeller_interactive(..., .mapping)
}
\arguments{
\item{...}{Arguments of the form \code{name = ~ expr} are used to create
\code{.mapping} (if \code{.mapping} is missing).  Other arguments (or all arguments
if \code{.mapping} is not missing) are passed through to \code{\link[ggplot2:labeller]{ggplot2::labeller()}}.}

\item{.mapping}{An aesthetic mapping as could be created with
\code{\link[ggplot2:aes]{ggplot2::aes()}} or \code{\link[ggplot2:aes_]{ggplot2::aes_()}}.  If missing  (the typical use case),
\code{.mapping} is created from the arguments in \code{...} that have the form
\code{name = ~ expr}.}
}
\value{
a labeller
}
\description{
Create interactive labeller
}
