% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_plot.R
\name{gf_plot}
\alias{gf_plot}
\title{Formula interface to ggplot()}
\usage{
gf_plot(...)
}
\arguments{
\item{...}{arguments that can include \code{data} (a data frame or something that can be
\code{\link[ggplot2:fortify]{ggplot2::fortify()}}ed to become one) and aesthetics specified using the following
formula notation: \code{aesthetic = ~ expression}.  See examples.}
}
\value{
a gg object
}
\description{
Create a new ggplot and (optionally) set default dataset aesthetics mapping.
}
\examples{
gf_plot(mtcars, x = ~ wt, y = ~ mpg, color = ~ factor(cyl)) \%>\%
  gf_density_2d() \%>\%
  gf_point()

}
