% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiraph-documentation-with-examples.R
\name{gf_point_interactive}
\alias{gf_point_interactive}
\title{Interactive scatter plots}
\usage{
gf_point_interactive(
  object = NULL,
  gformula = NULL,
  data = NULL,
  ...,
  alpha,
  color,
  size,
  shape,
  fill,
  group,
  stroke,
  xlab,
  ylab,
  title,
  subtitle,
  caption,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  show.help = NULL,
  inherit = TRUE,
  environment = parent.frame()
)
}
\arguments{
\item{object}{When chaining, this holds an object produced in the
earlier portions of the chain. Most users can safely ignore this
argument. See details and examples.}

\item{gformula}{A formula with shape \code{y ~ x}. Faceting can be
achieved by including \code{|} in the formula.}

\item{data}{A data frame with the variables to be plotted.}

\item{...}{Additional arguments passed to the underlying interactive
geom. This is where ggiraph's interactive aesthetics are supplied,
including \code{tooltip} (text shown on hover), \code{data_id}
(identifiers used for interactive selection), and \code{onclick}
(JavaScript run on click).}

\item{alpha}{Opacity (0 = invisible, 1 = opaque).}

\item{color}{A color or a formula used for mapping color.}

\item{size}{A numeric size or a formula used for mapping size.}

\item{shape}{An integer or letter shape or a formula used for mapping
shape.}

\item{fill}{A color for filling, or a formula used for mapping fill.}

\item{group}{Used for grouping.}

\item{stroke}{A numeric size of the border or a formula used to map
stroke.}

\item{xlab}{Label for x-axis. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{ylab}{Label for y-axis. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{title}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{subtitle}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{caption}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{stat}{A character string naming the stat used to make the layer.}

\item{position}{Either a character string naming the position function
used for the layer or a position object returned from a call to a
position function.}

\item{show.legend}{A logical indicating whether this layer should be
included in the legends. \code{NA}, the default, includes layer in the
legends if any of the attributes of the layer are mapped.}

\item{show.help}{If \code{TRUE}, display some minimal help.}

\item{inherit}{A logical indicating whether default attributes are
inherited.}

\item{environment}{An environment in which to look for variables not
found in \code{data}.}
}
\value{
A gg object that can be displayed with \code{\link[=gf_girafe]{gf_girafe()}}.
}
\description{
Creates an interactive plot using ggiraph. This function extends
\code{\link[=gf_point]{gf_point()}} with interactive features like tooltips and clickable elements.
}
\section{Additional interactive features}{

\itemize{
\item \code{onclick}: JavaScript code (as character string) executed when clicking elements.
\item Additional ggiraph aesthetics may be available depending on the geom.
}
}

\examples{
gf_point_interactive(mpg ~ wt, data = mtcars,
                    tooltip = ~ paste("Model:", rownames(mtcars))) |>
  gf_girafe()

# With color mapping and data_id for selection
gf_point_interactive(mpg ~ wt, data = mtcars,
                    color = ~ factor(cyl),
                    tooltip = ~ paste(rownames(mtcars), ":", mpg, "mpg"),
                    data_id = ~ rownames(mtcars)) |>
  gf_girafe()

}
\seealso{
\code{\link[=gf_point]{gf_point()}}, \code{\link[=gf_girafe]{gf_girafe()}}
}
