% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiraph-documentation-with-examples.R
\name{gf_rect_interactive}
\alias{gf_rect_interactive}
\title{Interactive rect plots}
\usage{
gf_rect_interactive(
  object = NULL,
  gformula = NULL,
  data = NULL,
  ...,
  alpha,
  color,
  fill,
  group,
  linetype,
  linewidth,
  xlab,
  ylab,
  title,
  subtitle,
  caption,
  stat = "identity",
  position = "identity",
  show.legend = NA,
  show.help = NULL,
  inherit = TRUE,
  environment = parent.frame()
)
}
\arguments{
\item{object}{When chaining, this holds an object produced in the
earlier portions of the chain. Most users can safely ignore this
argument. See details and examples.}

\item{gformula}{A formula with shape \code{ymin + ymax ~ xmin + xmax}.
Faceting can be achieved by including \code{|} in the formula.}

\item{data}{A data frame with the variables to be plotted.}

\item{...}{Additional arguments passed to the underlying interactive
geom. This is where ggiraph's interactive aesthetics are supplied,
including \code{tooltip} (text shown on hover), \code{data_id}
(identifiers used for interactive selection), and \code{onclick}
(JavaScript run on click).}

\item{alpha}{Opacity (0 = invisible, 1 = opaque).}

\item{color}{A color or a formula used for mapping color.}

\item{fill}{A color for filling, or a formula used for mapping fill.}

\item{group}{Used for grouping.}

\item{linetype}{A linetype (numeric or "dashed", "dotted", etc.) or a
formula used for mapping linetype.}

\item{linewidth}{A numerical line width or a formula used for mapping
linewidth.}

\item{xlab}{Label for x-axis. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{ylab}{Label for y-axis. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{title}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{subtitle}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{caption}{Title, sub-title, and caption for the plot. See also
\code{\link[=gf_labs]{gf_labs()}}.}

\item{stat}{A character string naming the stat used to make the layer.}

\item{position}{Either a character string naming the position function
used for the layer or a position object returned from a call to a
position function.}

\item{show.legend}{A logical indicating whether this layer should be
included in the legends. \code{NA}, the default, includes layer in the
legends if any of the attributes of the layer are mapped.}

\item{show.help}{If \code{TRUE}, display some minimal help.}

\item{inherit}{A logical indicating whether default attributes are
inherited.}

\item{environment}{An environment in which to look for variables not
found in \code{data}.}
}
\value{
A gg object that can be displayed with \code{\link[=gf_girafe]{gf_girafe()}}.
}
\description{
Creates an interactive plot using ggiraph. This function extends
\code{\link[=gf_rect]{gf_rect()}} with interactive features like tooltips and clickable elements.
}
\section{Additional interactive features}{

\itemize{
\item \code{onclick}: JavaScript code (as character string) executed when clicking elements.
\item Additional ggiraph aesthetics may be available depending on the geom.
}
}

\examples{
rect_data <-
  data.frame(
    x1 = c(1, 3, 1, 5, 4),
    x2 = c(2, 4, 3, 6, 6),
    y1 = c(1, 1, 4, 1, 3),
    y2 = c(2, 2, 5, 3, 5),
    t = c('a', 'a', 'a', 'b', 'b'),
    r = c(1, 2, 3, 4, 5),
    tooltip = c("ID 1", "ID 2", "ID 3", "ID 4", "ID 5"),
    uid = c("ID 1", "ID 2", "ID 3", "ID 4", "ID 5"),
    oc = rep("alert(this.getAttribute(\"data-id\"))", 5)
  )

p <- rect_data |>
  gf_rect_interactive(
    y1 + y2 ~ x1 + x2,
    fill = t,
    tooltip = ~ tooltip,
    onclick = ~ oc,
    data_id = ~ uid,
    color = "black",
    alpha = 0.5,
    linejoin = "bevel",
    lineend = "round"
  ) |>
  gf_text(
    (y1 + (y2 - y1) / 2) ~ (x1 + (x2 - x1) / 2),
    label = ~ r,
    size = 4
    )

if (interactive()) {
 p |> gf_girafe()
}



}
\seealso{
\code{\link[=gf_rect]{gf_rect()}}, \code{\link[=gf_girafe]{gf_girafe()}}
}
