% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{gf_relabel}
\alias{gf_relabel}
\title{Modify plot labeling}
\usage{
gf_relabel(plot, labels = get_labels(plot$data))
}
\arguments{
\item{plot}{A ggplot.}

\item{labels}{A named list of labels.}
}
\value{
A plot with potentially modified labels.
}
\description{
Some packages like expss provide mechanisms for providing longer labels to R objects.
These labels can be used when labeling plots and tables, for example, without requiring
long or awkward variable names.  This is an experimental feature and currently only supports
expss or any other system that stores a label in the \code{label} attribute of a vector.
}
\examples{

labels <- list(width = "width of foot (cm)", length = "length of foot (cm)",
  domhand = "dominant hand")
gf_point(length ~ width, color = ~domhand, data = mosaicData::KidsFeet) \%>\%
  gf_relabel(labels)

# Alternatively, we can store labels with data.
KF <- mosaicData::KidsFeet
attr(KF$length, 'label') = 'foot length (cm)'
attr(KF$width, 'label') = 'foot width (cm)'
gf_point(length ~ width, data = KF)
gf_density2d(length ~ width, data = KF)
get_labels(KF)


}
