% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_roxy.R
\name{gf_roxy}
\alias{gf_roxy}
\title{functions.
For plots with just one layer, the formula interface
is more compact and is consistent with modeling and \\pkg{{mosaic}} notation.
The functions generate a \\code{{ggplot}} command string which can be displayed by
setting \\code{{verbose = TRUE}} as an argument.}
\usage{
gf_roxy(type, aes_form = y ~ x, extras = alist(), note = NULL)
}
\description{
Positional aesthetics are typically specified using a formula (see the \\code{{gformula}} argument).
Additional formula terms of the form \\code{{+ attribute::value}} map \\code{{attribute}}
to \\code{{value}}.
Additional terms of the form \\code{{+ attribute:value}} will map \\code{{attribute}}
to \\code{{value}} if \\code{{value}} is the name of a variable in \\code{{data}}, else
\\code{{attribute}} will be set to the constant \\code{{value}}. Alternatively (and preferably)
attributes can be set can be set using arguments of the form \\code{{attribute = value}} or
mapped using arguments of the form \\code{{attribute = ~ expression}}.",
\\code{{\\link{{facet_wrap}}()}} or \\code{{\\link{{facet_grid}}()}}.
This provides an alternative to
\\code{{\\link{{gf_facet_wrap}}()}} and
\\code{{\\link{{gf_facet_grid}}()}} that is terser and may feel more familiar to users
of \\pkg{{lattice}}.
}
\details{
Evaluation of the \\pkg{{ggplot2}} code occurs in the environment of \\code{{gformula}}.
This will typically do the right thing when formulas are created on the fly, but might not
be the right thing if formulas created in one environment are used to create plots
in another.",
of the chain.  Most users can safely ignore this argument.
See details and examples.
",
  Faceting can be acheived by including \\code{{|}} in the formula.",
  (a) ggplot2 aesthetics to be set with \\code{{attribute = value}},
  (b) ggplot2 aesthetics to be mapped with \\code{{attribute = ~expression}}, or
  (c) attributes of the layer as a whole, which are set with \\code{{attribute = value}}.
  Available attributes include ",
  can be added to an existing frame.",
}
