% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_aux.R
\name{gf_theme}
\alias{gf_theme}
\title{Themes for ggformula}
\usage{
gf_theme(object, theme, ...)
}
\arguments{
\item{object}{a gg object}

\item{theme}{a ggplot2 theme function like \code{\link[ggplot2:theme_minimal]{ggplot2::theme_minimal()}}.}

\item{...}{If \code{theme} is missing, then these additional arguments are theme elements
of the sort handled by \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
a modified gg object
}
\description{
Themes for ggformula
}
