% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newplots.R
\name{StatAsh}
\alias{StatAsh}
\alias{StatSpline}
\alias{StatQqline}
\alias{StatLm}
\alias{GeomLm}
\alias{StatFitdistr}
\title{ggproto classes for ggplot2}
\usage{
StatAsh

StatSpline

StatQqline

StatLm

GeomLm

StatAsh

StatFitdistr
}
\description{
These are typically accessed through their associated \verb{geom_*}, \verb{stat_*} or
\verb{gf_*} functions.

These are typically accessed through their associated \verb{geom_*}, \verb{stat_*} or
\verb{gf_*} functions.
}
\seealso{
\code{\link[=stat_ash]{stat_ash()}}

\code{\link[=gf_ash]{gf_ash()}}

\code{\link[=stat_spline]{stat_spline()}}

\code{\link[=gf_spline]{gf_spline()}}

\code{\link[ggplot2:stat_qq]{ggplot2::stat_qq()}}

\code{\link[=gf_qq]{gf_qq()}}

\code{\link[=stat_lm]{stat_lm()}}

\code{\link[=gf_lm]{gf_lm()}}

\code{\link[=geom_lm]{geom_lm()}}
}
