% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstance-deprecated.R
\name{gf_barh}
\alias{gf_barh}
\alias{ggstance}
\alias{gf_countsh}
\alias{gf_colh}
\alias{gf_propsh}
\alias{gf_percentsh}
\alias{gf_boxploth}
\alias{gf_linerangeh}
\alias{gf_pointrangeh}
\alias{gf_crossbarh}
\alias{gf_violinh}
\alias{gf_errorbarh}
\title{Deprecated horizontal plotting functions}
\usage{
gf_barh(...)

gf_countsh(...)

gf_colh(...)

gf_propsh(...)

gf_percentsh(...)

gf_boxploth(...)

gf_linerangeh(...)

gf_pointrangeh(...)

gf_crossbarh(...)

gf_violinh(...)

gf_errorbarh(...)
}
\arguments{
\item{...}{additional arguments}
}
\description{
These functions were wrappers around functions from \code{ggstance} from an era
before \code{ggplot2} supported horizonally oriented geoms.  \code{ggstance} has not
been updated to comply with the current version of \code{ggplot2}, and since the
functionalilty is now available by other means, these functions have been
deprecated.
}
\examples{
gf_violin(carat ~ color, data = diamonds)
gf_violin(carat ~ color, data = diamonds) |>
  gf_refine(coord_flip())
gf_violin(color ~ carat, data = diamonds)
gf_density(~ carat, data = diamonds)
gf_density(carat ~ ., data = diamonds)
}
