% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggiraph.R
\name{interactive_facets}
\alias{interactive_facets}
\alias{gf_facet_wrap_interactive}
\alias{gf_facet_grid_interactive}
\title{Interactive facets}
\usage{
gf_facet_wrap_interactive(
  object,
  ...,
  labeller,
  interactive_on = c("text", "rect", "both")
)

gf_facet_grid_interactive(
  object,
  ...,
  labeller,
  interactive_on = c("text", "rect", "both")
)
}
\arguments{
\item{object}{a ggplot graphic}

\item{...}{additional arguments passed to \code{labeller} and to the
ggplot2 faceting function (\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} or \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}).}

\item{labeller}{a labeller created using \code{\link[=gf_labeller_interactive]{gf_labeller_interactive()}}}

\item{interactive_on}{one of "text" (strip text is made interactive),
"rect" (strip rectangles are made interactive), or "both". Can be abbreviated.}
}
\description{
To create interactive facets, use \code{gf_facet_wrap_interactive()} or
\code{gf_facet_grid_interactive()} and use \code{\link[=gf_labeller_interactive]{gf_labeller_interactive()}}
to create the \code{labeller}.
}
\seealso{
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}

\code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}

\code{\link[=gf_labeller_interactive]{gf_labeller_interactive()}}
}
