% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIpop-doc.R
\docType{data}
\name{MIpop}
\alias{MIpop}
\title{Population of Michigan counties}
\format{A data frame with populations of Michigan counties.
\describe{
\item{rank}{ Population rank.}
\item{county}{ County name.}
\item{population}{ Population (2010 census).}
}}
\usage{
data(MIpop)
}
\description{
Population of Michigan counties
}
