% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newplots.R
\docType{data}
\name{qq.line}
\alias{qq.line}
\alias{StatQqline}
\alias{stat_qqline}
\title{A stat for adding guide lines to qq-plots}
\format{An object of class \code{StatQqline} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
qq.line(sample, qdist, na.rm = TRUE, tail = 0.25)

StatQqline

stat_qqline(mapping = NULL, data = NULL, geom = "line",
  position = "identity", ..., distribution = stats::qnorm,
  dparams = list(), na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{na.rm}{A logical indicating whether a warning should be issued when
missing values are removed before plotting.}

\item{tail}{A tail probability.  The constructed line will connect the \code{tail} and \code{1 - tail} quantiles
of the sample and theoretical distributions.}

\item{mapping}{An aesthetic mapping produced with \code{\link{aes}()} or
\code{\link{aes_string}()}.}

\item{data}{A data frame.}

\item{geom}{A geom.}

\item{position}{A position object.}

\item{...}{Additional arguments}

\item{show.legend}{A logical indicating whether legends should be included
for this layer.  If \code{NA}, legends will be inclued for each aesthetic
that is mapped.}

\item{inherit.aes}{A logical indicating whether aesthetics should be
inherited.  When \code{FALSE}, the supplied \code{mapping} will be
the only aesthetics used.}

\item{stat}{A stat.}
}
\description{
This stat computes points on a line connecting two quantiles
of the sample and theoretical distributions.
}
\examples{
ggplot(data = iris, aes(sample = Sepal.Length)) +
  geom_qq() +
  stat_qqline( alpha = 0.7, color = "red", linetype = "dashed") +
  facet_wrap(~Species)

}
\keyword{datasets}
