% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel.R
\name{get_labels}
\alias{get_labels}
\alias{print.gf_ggplot}
\title{Extract labels from a labeled object}
\usage{
get_labels(object, ...)

\method{print}{gf_ggplot}(x, labels = get_labels(x$data), ...)
}
\arguments{
\item{object}{An R object, potentially labelled.}

\item{...}{Additional arguments. Ignored in existing methods.}

\item{x}{A ggplot.}
}
\value{
The label(s) for the object.  If \code{object} is a data frame,
labels are extracted from each variable and returned as a named list.
If there are no labels, \code{NULL} is returned.
}
\description{
Some packages like expss provide mechanisms for providing longer labels to R objects.
These labels can be used when labeling plots and tables, for example, without requiring
long or awkward variable names.  This is an experimental feature and currently only supports
expss or any other system that stores a label in the \code{label} attribute of a vector.
}
\examples{
KF <- mosaicData::KidsFeet
attr(KF$length, 'label') = 'foot length (cm)'
attr(KF$width, 'label') = 'foot width (cm)'
gf_point(length ~ width, data = KF)
get_labels(KF)


}
