% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_functions.R
\name{gf_fitdistr}
\alias{gf_fitdistr}
\title{Plot density function based on fit to data}
\usage{
gf_fitdistr(
  object = NULL,
  gformula = NULL,
  data = NULL,
  ...,
  dist = "dnorm",
  start = NULL,
  alpha,
  color,
  fill,
  group,
  linetype,
  linewidth,
  size,
  xlab,
  ylab,
  title,
  subtitle,
  caption,
  geom = "path",
  stat = "fitdistr",
  position = "identity",
  show.legend = NA,
  show.help = NULL,
  inherit = FALSE,
  environment = parent.frame()
)
}
\arguments{
\item{object}{When chaining, this holds an object produced in the earlier portions
of the chain.  Most users can safely ignore this argument. See examples.}

\item{gformula}{A formula with shape \code{ ~ x} used to specify the data
to be fit to a family of distributions.}

\item{data}{A data frame containing the variable to be fitted.}

\item{...}{Additional arguments}

\item{dist}{A quoted name of a distribution function.
See \code{\link[mosaicCore:fit_distr_fun]{mosaicCore::fit_distr_fun()}} for more details about allowable distributions.}

\item{start}{Starting value(s) for the search for MLE.  (See \link[MASS:fitdistr]{MASS::fitdistr}.)}

\item{alpha}{Opacity (0 = invisible, 1 = opaque).}

\item{color}{A color or a formula used for mapping color.}

\item{fill}{A color for filling, or a formula used for mapping fill.}

\item{group}{Used for grouping.}

\item{linetype}{A linetype (numeric or "dashed", "dotted", etc.) or a formula used
for mapping linetype.}

\item{linewidth}{A numerical line width or a formula used for mapping linewidth.}

\item{size}{size aesthetic for dots in pmf plots.}

\item{xlab}{Label for x-axis. See also \code{\link[=gf_labs]{gf_labs()}}.}

\item{ylab}{Label for y-axis. See also \code{\link[=gf_labs]{gf_labs()}}.}

\item{title, subtitle, caption}{Title, sub-title, and caption for the plot.
See also \code{\link[=gf_labs]{gf_labs()}}.}

\item{geom}{A character string naming the geom used to make the layer.}

\item{stat}{A character string naming the stat used to make the layer.}

\item{position}{Either a character string naming the position function used
for the layer or a position object returned from a call to a position function.}

\item{show.legend}{A logical indicating whether this layer should be included in
the legends.  \code{NA}, the default, includes layer in the legends if any
of the attributes of the layer are mapped.}

\item{show.help}{If \code{TRUE}, display some minimal help.}

\item{inherit}{A logical indicating whether default attributes are inherited.}

\item{environment}{An environment in which to look for variables not found in \code{data}.}
}
\value{
a gg object
}
\description{
\code{MASS::fitdistr()} is used to fit coefficients of a specified family of
distributions and the resulting density curve is displayed.
}
\section{Specifying plot attributes}{


Positional attributes (a.k.a, aesthetics) are specified using the formula in \code{gformula}.
Setting and mapping of additional attributes can be done through the
use of additional arguments.
Attributes can be set can be set using arguments of the form \code{attribute = value} or
mapped using arguments of the form \code{attribute = ~ expression}.

In formulas of the form \code{A | B}, \code{B} will be used to form facets using
\code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} or \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}.
This provides an alternative to
\code{\link[=gf_facet_wrap]{gf_facet_wrap()}} and
\code{\link[=gf_facet_grid]{gf_facet_grid()}} that is terser and may feel more familiar to users
of \pkg{lattice}.
}

\section{Evaluation}{


Evaluation of the \pkg{ggplot2} code occurs in the environment of \code{gformula}.
This will typically do the right thing when formulas are created on the fly, but might not
be the right thing if formulas created in one environment are used to create plots
in another.
}

\examples{
gf_fitdistr(~length, data = mosaicData::KidsFeet, inherit = FALSE) |>
  gf_dhistogram(~length, data = mosaicData::KidsFeet, binwidth = 0.5, alpha = 0.25)

gf_dhistogram(~length, data = mosaicData::KidsFeet, binwidth = 0.5, alpha = 0.25) |>
  gf_fitdistr()

set.seed(12345)
Dat <- data.frame(
  f = rf(500, df1 = 3, df2 = 47),
  g = rgamma(500, 3, 10)
)
gf_dhistogram(~g, data = Dat) |>
  gf_fitdistr(dist = "dgamma", linewidth = 1.4)
fitted_density <- mosaicCore::fit_distr_fun(~g, data = Dat, dist = "dgamma")
gf_dhistogram(~g, data = Dat) |>
  gf_fun(fitted_density(x) ~ x, inherit = FALSE)

gf_dhistogram(~f, data = Dat) |>
  gf_fitdistr(dist = "df", start = list(df1 = 2, df2 = 50))

# fitted parameters are default argument values
args(
  mosaicCore::fit_distr_fun(~f,
    data = Dat, dist = "df",
    start = list(df1 = 2, df2 = 50)
  )
)
args(mosaicCore::fit_distr_fun(~g, data = Dat, dist = "dgamma"))
}
\seealso{
\code{\link[mosaicCore:fit_distr_fun]{mosaicCore::fit_distr_fun()}}
}
