% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf_functions.R
\name{gf_histogram}
\alias{gf_histogram}
\alias{gf_density}
\alias{gf_dens}
\alias{gf_dotplot}
\alias{gf_bar}
\alias{gf_freqpoly}
\alias{gf_qq}
\alias{gf_rug}
\title{Univariate gf_ plotting functions}
\usage{
gf_histogram(object = NULL, gformula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(object, c("gg", "ggplot")), ...,
  position = NULL, show.help = is.null(object) && is.null(gformula))

gf_density(object = NULL, gformula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(object, c("gg", "ggplot")), ...,
  position = NULL, show.help = is.null(object) && is.null(gformula))

gf_dens(object = NULL, gformula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(object, c("gg", "ggplot")), ...,
  position = NULL, show.help = is.null(object) && is.null(gformula))

gf_dotplot(object = NULL, gformula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(object, c("gg", "ggplot")), ...,
  position = NULL, show.help = is.null(object) && is.null(gformula))

gf_bar(object = NULL, gformula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(object, c("gg", "ggplot")), ...,
  position = NULL, show.help = is.null(object) && is.null(gformula))

gf_freqpoly(object = NULL, gformula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(object, c("gg", "ggplot")), ...,
  position = NULL, show.help = is.null(object) && is.null(gformula))

gf_qq(object = NULL, gformula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(object, c("gg", "ggplot")), ...,
  position = NULL, show.help = is.null(object) && is.null(gformula))

gf_rug(object = NULL, gformula = NULL, data = NULL, geom = type,
  verbose = FALSE, add = inherits(object, c("gg", "ggplot")), ...,
  position = NULL, show.help = is.null(object) && is.null(gformula))
}
\arguments{
\item{object}{When chaining, this holds an object produced in the earlier portions
of the chain.  Most users can safely ignore this argument.
See details and examples.}

\item{gformula}{A formula describing the x variable and other aesthetics in
a form like \code{ ~ x + color:red + fill:gray50 + alpha:0.5}.
The environment of \code{gformula} determines
where the \pkg{ggplot2} code is evaluated. See details.}

\item{data}{A data frame with the variables to be plotted}

\item{geom}{A way to specify ggplot geoms that are not aliased to gf functions.}

\item{verbose}{If \code{TRUE} print the ggplot2 command in the console.}

\item{add}{If \code{TRUE} then construct just the layer with no frame.  The result
can be added to an existing frame.}

\item{...}{Other arguments such as \code{position="dodge"}.}

\item{show.help}{If \code{TRUE}, display some minimal help.  In particular,
the help will show (a) which geom from \pkg{ggplot2} is used,
(b) how aesthetics are assigned based on \code{gformula}, and (c)
any default values of arguments to the geom.}
}
\description{
These functions provide a formula interface to \code{ggplot2} and
various geoms. The formula interface is similar to the one used
for \pkg{lattice} plots, but more expressive, and consistent with
its use in modeling functions like \code{\link{lm}()}.  These functions
can be used to create a complete plot, or they can be chained together
using the pipe operator from \pkg{magrittr} to create multi-layer plots.
The functions generate a \code{ggplot2} command string which can be
displayed by setting \code{verbose = TRUE} as an argument.
}
\details{
Formulas must specify the \code{x} aesthetic in the form \code{~x}.
Additional terms of the form \code{attribute::value} map \code{attribute}
to \code{value}.
Additional terms of the form \code{attribute:value} will map \code{attribute}
to \code{value} if \code{value} is the name of a variable in \code{data}, else
\code{attribute} will be set to the constant \code{value}.
Attributes can also be set by including optional arguments of the form
\code{attribute = value}.

In formulas of the form \code{A | B}, \code{B} will be used to form facets using
\code{link{facet_wrap}()} or \code{\link{facet_grid}()}.
This provides an alternative to
\code{\link{gf_facet_wrap}()} and
\code{\link{gf_facet_grid}()} that is terser and may feel more familiar to users
of \pkg{lattice}.

Evaluation of the \pkg{ggplot2} code occurs in the environment of \code{formula}.  This
will typically do the right thing when formulas are created on the fly, but might not
be the right thing if formulas created in one environment are used to create plots
in another.
}
\examples{
gf_dens(show.help = TRUE)
gf_histogram(~ Sepal.Length | Species, data = iris, binwidth = 0.25)
gf_density(~ Sepal.Length + color:Species, data = iris)
gf_dens(~ Sepal.Length + color:Species, data = iris)
gf_freqpoly(~ Sepal.Length + color:Species, data = iris)
gf_dotplot(~ Sepal.Length + fill:Species, data = iris)
# Chaining in the data
iris \%>\% gf_dens(~ Sepal.Length + color:Species)
}
\seealso{
\code{\link{gf_point}()}, \code{\link{gf_abline}()}, \code{\link{gf_pointrange}()}, \code{\link{gf_refine}()}, and the other functions documented with these functions.
}
