% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newplots.R
\docType{data}
\name{StatAsh}
\alias{StatAsh}
\alias{StatSpline}
\alias{StatQqline}
\alias{StatLm}
\alias{GeomLm}
\title{ggproto classes for ggplot2}
\usage{
StatAsh

StatSpline

StatQqline

StatLm

GeomLm
}
\description{
These are typically accessed through their associated \code{geom_*}, \code{stat_*} or
\code{gf_*} functions.
}
\seealso{
\code{\link[=stat_ash]{stat_ash()}}

\code{\link[=gf_ash]{gf_ash()}}

\code{\link[=stat_spline]{stat_spline()}}

\code{\link[=gf_spline]{gf_spline()}}

\code{\link[=stat_qq]{stat_qq()}}

\code{\link[=gf_qq]{gf_qq()}}

\code{\link[=stat_lm]{stat_lm()}}

\code{\link[=gf_lm]{gf_lm()}}

\code{\link[=geom_lm]{geom_lm()}}

\code{\link[=gf_lm]{gf_lm()}}
}
\keyword{datasets}
