% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newplots.R
\docType{data}
\name{StatAsh}
\alias{StatAsh}
\alias{StatSpline}
\alias{StatQqline}
\alias{StatLm}
\alias{GeomLm}
\title{Geoms and stats for spline smoother}
\format{An object of class \code{StatAsh} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
StatAsh

StatSpline

StatQqline
}
\description{
Functions to allow spline smoothing with ggplot2

This stat computes points on a line connecting two quantiles
of the sample and theoretical distributions.
}
\keyword{datasets}
