% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_factory.R
\name{interactive_layer_factory}
\alias{interactive_layer_factory}
\title{Create an interactive ggformula layer function}
\usage{
interactive_layer_factory(geom_fun)
}
\arguments{
\item{geom_fun}{A character string naming an interactive geom (example: "geom_point_interactive")}
}
\description{
Primarily intended for package developers, this function factory
is used to create layer functions in the ggformula package.
}
