% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{makeFun}
\alias{df_stats}
\alias{na.warn}
\alias{var_label}
\alias{var_label<-}
\alias{set_variable_labels}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{labelled}{\code{\link[labelled:var_label]{set_variable_labels}}, \code{\link[labelled]{var_label}}, \code{\link[labelled:var_label]{var_label<-}}}

  \item{mosaicCore}{\code{\link[mosaicCore]{df_stats}}, \code{\link[mosaicCore]{makeFun}}, \code{\link[mosaicCore]{na.warn}}}
}}

