% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{makeFun}
\alias{df_stats}
\alias{na.warn}
\alias{stat_sina}
\alias{StatSina}
\alias{geom_sina}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggforce}{\code{\link[ggforce]{geom_sina}}, \code{\link[ggforce:geom_sina]{stat_sina}}, \code{\link[ggforce:ggforce-extensions]{StatSina}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{mosaicCore}{\code{\link[mosaicCore]{df_stats}}, \code{\link[mosaicCore]{makeFun}}, \code{\link[mosaicCore]{na.warn}}}
}}

